requirements.txt
aioredis==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
cssselect2==0.7.0
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.1.14
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-classy-tags==3.0.1
django-cms==3.8.0
django-filer==2.2.3
django-formtools==2.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-timezone-field==5.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.1.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.14.0
easy-thumbnails==2.8.3
entrypoints==0.4
executing==1.1.1
fake-useragent==0.1.11
feedparser==6.0.10
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kombu==5.2.4
lxml==4.9.1
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg==3.1.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
redis==4.3.4
reportlab==3.6.11
requests==2.28.1
requests-html==0.10.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
svglib==1.4.1
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
Unidecode==1.1.2
urllib3==1.26.12
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
wrapt==1.14.1
yahoo-fin==0.8.9.1
zipp==3.8.1
zope.interface==5.4.0
